Implement RNAProductAdaptor::list_dbIDs()
At this point the only reason for the current set of tests still failing
should be - and unless I have misread the output, it is - that there is
still no rnaproduct table in the database. Will get to that bit shortly.

Nb. it uses the same kind of reimplementation of the internal function
_list_dbIDs() as TranslationAdaptor so that multi-species scenarios
are handled correctly.