RNAProduct: set default value for stable_id in the constructor
If $stable_id is undef, a discrepancy exists in values returned by
stable_id() and stable_id_version(): the former returns an undef, the
latter returns "." or ".$version" (implying the stable-id part to be an
empty string) + throws a warning about using an undefined value in
concatenation. Have the constructor set $stable_id to an empty string if
no value has been passed so that we both get consistent answers and
quell the warning.