RNAProductAdaptor: tell perlcritic not to complain about _tables()
By default, at severity 3 or higher perlcritic complains about _tables()
having been declared but not being used. In reality this method *is*
used but in the superclass Bio::EnsEMBL::DBSQL::BaseAdaptor (indeed, it
is declared as abstract over there so the code will not even work
without a concrete implementation in the subclass), unfortunately
perlcritic doesn't handle this. Therefore, we now tell it to relax this
policy for this specific subroutine.