RNAProduct::get_all_Attributes(): do not throw when data unavailable
Previously, get_all_Attributes() would throw an exception in the event
of absence of both the local attribute cache and the adaptor link. That
was not entirely correct because a locally created RNAProduct object
which has not been assigned any attributes would look exactly like that.
get_all_DBEntries() got that right so let's follow its example.