id_mapping: implement rnaproduct mapping in InternalIdMapper
Mostly copied from map_translation(), which is why it is in the main
package and not a separate plug-in. On the other hand, a transcript can
have multiple mature RNA products associated with it so the comparison
between source and target rnaproducts is more complicated than for
translations. In particular, how exactly this comparison works depends
on the type of RNA products in question. Yes, that does pretty much mean
we will not be able to assign stable IDs to generic RNAProducts and
might reconsider whether to allow writing them to the database in the
first place.

For MicroRNAs, which at the moment are the only specific type of mature
RNA the Core API supports, we look at what arm of the precursor hairpin
they have come and create a mapping for each source-target pair coming
from the same arm.

Given this will grow as new rnaproduct types are added to Ensembl, we
might want to implement this as a plug-in after all.